refactor: deduplicate shared widgets and merge identical callbacks
Desktop:
- Merge identical handleMouseLeave/handleBlur into single scheduleClose

Mobile:
- Extract SmallAvatar to shared small_avatar.dart
- Move shortPubkey and formatMessageTime to date_formatters.dart
- Remove duplicated _showThreadMessageActions in favor of shared showMessageActions
- Remove unused import and stale comments
